Move ChunkParser into a seperate file
1. Move ChunkParser class into chunkparser.py
2. Amend tests for ChunkParser to use unittest framework
3. Fix commented out symmetry.
4. Fix incorrect fixed planes being applied.
